+++ explcheck/src/explcheck-semantic-analysis.lua
@@ -1,6 +1,5 @@
 -- The semantic analysis step of static analysis determines the meaning of the different function calls.
 
-local get_option = require("explcheck-config").get_option
 local token_types = require("explcheck-lexical-analysis").token_types
 local syntactic_analysis = require("explcheck-syntactic-analysis")
 local ranges = require("explcheck-ranges")
@@ -59,6 +58,14 @@ local DEFINITELY = statement_confidences.DEFINITELY
 local MAYBE = statement_confidences.MAYBE
 local NONE = statement_confidences.NONE
 
+local csname_types = {
+  TEXT = "direct text representation of a control sequence name or its part, usually paired with confidence DEFINITELY",
+  PATTERN = "a PEG pattern that recognizes different control sequences or their parts, usually paired with confidence MAYBE"
+}
+
+local TEXT = csname_types.TEXT
+local PATTERN = csname_types.PATTERN
+
 local simple_text_catcodes = {
   [3] = true,  -- math shift
   [4] = true,  -- alignment tab
@@ -72,7 +79,7 @@ local simple_text_catcodes = {
 }
 
 -- Determine the meaning of function calls and register any issues.
-local function semantic_analysis(pathname, content, issues, results, options)
+local function semantic_analysis(pathname, content, issues, results, options)  -- luacheck: ignore pathname options
 
   -- Determine the type of a span of tokens as either "simple text" [1, p. 383] with no expected side effects or
   -- a more complex material that may have side effects and presents a boundary between chunks of well-understood
@@ -123,32 +130,132 @@ local function semantic_analysis(pathname, content, issues, results, options)
         return csname_stem, argument_specifiers
       end
 
+      -- Determine whether a function is private or public based on its name.
+      local function is_function_private(csname)
+        return csname:sub(1, 2) == "__"
+      end
+
       -- Replace the argument specifiers in an expl3 control sequence name.
-      local function replace_argument_specifiers(csname, argument_specifiers)
-        local csname_stem, base_argument_specifiers = parse_expl3_csname(csname)
-        if csname_stem == nil or base_argument_specifiers == nil then
-          return nil  -- we couldn't parse the csname, give up
+      local function replace_argument_specifiers(csname_stem, argument_specifiers)
+        local csname
+        if type(argument_specifiers) == 'string' then
+          csname = string.format("%s:%s", csname_stem, argument_specifiers)
+        else
+          local transcript = string.format("%s:%s", csname_stem, argument_specifiers.transcript)
+          if argument_specifiers.type == TEXT then
+            csname = {
+              payload = string.format("%s:%s", csname_stem, argument_specifiers.payload),
+              transcript = transcript,
+              type = TEXT
+            }
+          elseif argument_specifiers.type == PATTERN then
+            csname = {
+              payload = lpeg.P(csname_stem) * lpeg.P(":") * argument_specifiers.payload,
+              transcript = transcript,
+              type = PATTERN
+            }
+          else
+            error('Unexpected argument specifiers type "' .. argument_specifiers.type .. '"')
+          end
         end
-        return string.format("%s:%s", csname_stem, argument_specifiers)
+        return csname
       end
 
       -- Determine the control sequence name of a conditional function given a base control sequence name and a condition.
-      local function get_conditional_function_csname(csname, condition)
-        local csname_stem, argument_specifiers = parse_expl3_csname(csname)
-        if csname_stem == nil or argument_specifiers == nil then
-          return nil  -- we couldn't parse the csname, give up
-        end
+      local function get_conditional_function_csname(csname_stem, argument_specifiers, condition)
+        local csname
         if condition == "p" then  -- predicate function
-          return string.format("%s_p:%s", csname_stem, argument_specifiers)
+          local format = "%s_p:%s"
+          if type(argument_specifiers) == 'string' then
+            csname = string.format(format, csname_stem, argument_specifiers)
+          else
+            local transcript = string.format(format, csname_stem, argument_specifiers.transcript)
+            if argument_specifiers.type == TEXT then
+              csname = {
+                payload = string.format(format, csname_stem, argument_specifiers.payload),
+                transcript = transcript,
+                type = TEXT
+              }
+            elseif argument_specifiers.type == PATTERN then
+              csname = {
+                payload = lpeg.P(csname_stem) * lpeg.P("_p:") * argument_specifiers.payload,
+                transcript = transcript,
+                type = PATTERN
+              }
+            else
+              error('Unexpected argument specifiers type "' .. argument_specifiers.type .. '"')
+            end
+          end
         elseif condition == "T" then  -- true-branch conditional function
-          return string.format("%s:%sT", csname_stem, argument_specifiers)
+          local format = "%s:%sT"
+          if type(argument_specifiers) == 'string' then
+            csname = string.format(format, csname_stem, argument_specifiers)
+          else
+            local transcript = string.format(format, csname_stem, argument_specifiers.transcript)
+            if argument_specifiers.type == TEXT then
+              csname = {
+                payload = string.format(format, csname_stem, argument_specifiers.payload),
+                transcript = transcript,
+                type = TEXT
+              }
+            elseif argument_specifiers.type == PATTERN then
+              csname = {
+                payload = lpeg.P(csname_stem) * lpeg.P(":") * argument_specifiers.payload * lpeg.P("T"),
+                transcript = transcript,
+                type = PATTERN
+              }
+            else
+              error('Unexpected argument specifiers type "' .. argument_specifiers.type .. '"')
+            end
+          end
         elseif condition == "F" then  -- false-branch conditional function
-          return string.format("%s:%sF", csname_stem, argument_specifiers)
+          local format = "%s:%sF"
+          if type(argument_specifiers) == 'string' then
+            csname = string.format(format, csname_stem, argument_specifiers)
+          else
+            local transcript = string.format(format, csname_stem, argument_specifiers.transcript)
+            if argument_specifiers.type == TEXT then
+              csname = {
+                payload = string.format(format, csname_stem, argument_specifiers.payload),
+                transcript = transcript,
+                type = TEXT
+              }
+            elseif argument_specifiers.type == PATTERN then
+              csname = {
+                payload = lpeg.P(csname_stem) * lpeg.P(":") * argument_specifiers.payload * lpeg.P("F"),
+                transcript = transcript,
+                type = PATTERN
+              }
+            else
+              error('Unexpected argument specifiers type "' .. argument_specifiers.type .. '"')
+            end
+          end
         elseif condition == "TF" then  -- true-and-false-branch conditional function
-          return string.format("%s:%sTF", csname_stem, argument_specifiers)
+          local format = "%s:%sTF"
+          if type(argument_specifiers) == 'string' then
+            csname = string.format(format, csname_stem, argument_specifiers)
+          else
+            local transcript = string.format(format, csname_stem, argument_specifiers.transcript)
+            if argument_specifiers.type == TEXT then
+              csname = {
+                payload = string.format(format, csname_stem, argument_specifiers.payload),
+                transcript = transcript,
+                type = TEXT
+              }
+            elseif argument_specifiers.type == PATTERN then
+              csname = {
+                payload = lpeg.P(csname_stem) * lpeg.P(":") * argument_specifiers.payload * lpeg.P("TF"),
+                transcript = transcript,
+                type = PATTERN,
+              }
+            else
+              error('Unexpected argument specifiers type "' .. argument_specifiers.type .. '"')
+            end
+          end
         else
           error('Unexpected condition "' .. condition .. '"')
         end
+        return csname
       end
 
       -- Try and extract a list of conditions in a conditional function (variant) definition.
@@ -251,32 +358,34 @@ local function semantic_analysis(pathname, content, issues, results, options)
             end
             ::continue::
           end
-          table.insert(variant_argument_specifiers, {argument_specifiers, DEFINITELY})
+          table.insert(variant_argument_specifiers, {
+            payload = argument_specifiers,
+            transcript = argument_specifiers,
+            type = TEXT,
+            confidence = DEFINITELY
+          })
         end
         goto done_parsing
 
         ::unknown_argument_specifiers::
         -- assume all possible sets of variant argument specifiers with lower confidence
         do
-          local max_intermediate_result_size = get_option("max_guessed_specifiers", options, pathname)
-          variant_argument_specifiers_list = {""}
+          variant_argument_specifiers = {}
+          local compatible_specifier_pattern, compatible_specifier_transcripts = parsers.success, {}
           for i = 1, #base_argument_specifiers do
             local base_argument_specifier = base_argument_specifiers:sub(i, i)
-            local intermediate_result = {}
-            for _, prefix in ipairs(variant_argument_specifiers_list) do
-              for _, compatible_specifier in ipairs(lpeg.match(parsers.compatible_argument_specifiers, base_argument_specifier)) do
-                table.insert(intermediate_result, prefix .. compatible_specifier)
-              end
-            end
-            if #intermediate_result > max_intermediate_result_size then  -- if there are too many possible variant argument specifiers
-              return nil  -- give up to prevent a combinatorial explosion
-              -- TODO: produce a special "wildcard" return value instead; this will complicate processing but should remain O(1)
-            end
+            local compatible_specifiers = table.concat(lpeg.match(parsers.compatible_argument_specifiers, base_argument_specifier))
+            compatible_specifier_pattern = compatible_specifier_pattern * lpeg.S(compatible_specifiers)
+            local compatible_specifier_transcript = string.format('[%s]', compatible_specifiers)
+            table.insert(compatible_specifier_transcripts, compatible_specifier_transcript)
           end
-        end
-        variant_argument_specifiers = {}
-        for _, argument_specifiers in ipairs(variant_argument_specifiers_list) do
-          table.insert(variant_argument_specifiers, {argument_specifiers, MAYBE})
+          local compatible_specifiers_transcript = table.concat(compatible_specifier_transcripts)
+          table.insert(variant_argument_specifiers, {
+            payload = compatible_specifier_pattern,
+            transcript = compatible_specifiers_transcript,
+            type = PATTERN,
+            confidence = MAYBE
+          })
         end
 
         ::done_parsing::
@@ -308,6 +417,10 @@ local function semantic_analysis(pathname, content, issues, results, options)
             goto other_statement
           end
           assert(base_csname ~= nil)
+          local base_csname_stem, base_argument_specifiers = parse_expl3_csname(base_csname)
+          if base_csname_stem == nil then  -- we couldn't parse the csname, give up
+            goto other_statement
+          end
           -- determine the variant argument specifiers
           local variant_argument_specifiers = parse_variant_argument_specifiers(base_csname, call.arguments[2])
           if variant_argument_specifiers == nil then  -- we couldn't parse the variant argument specifiers, give up
@@ -315,12 +428,7 @@ local function semantic_analysis(pathname, content, issues, results, options)
           end
           -- determine all defined csnames
           local defined_csnames = {}
-          for _, argument_specifier_table in ipairs(variant_argument_specifiers) do
-            local argument_specifiers, argument_specifier_confidence = table.unpack(argument_specifier_table)
-            local defined_csname = replace_argument_specifiers(base_csname, argument_specifiers)
-            if defined_csname == nil then  -- we couldn't determine the defined csname, give up
-              goto other_statement
-            end
+          for _, argument_specifiers in ipairs(variant_argument_specifiers) do
             if is_conditional then  -- conditional function
               -- determine the conditions
               local conditions = parse_conditions(call.arguments[#call.arguments])
@@ -330,19 +438,17 @@ local function semantic_analysis(pathname, content, issues, results, options)
               -- determine the defined csnames
               for _, condition_table in ipairs(conditions) do
                 local condition, condition_confidence = table.unpack(condition_table)
-                local base_conditional_csname = get_conditional_function_csname(base_csname, condition)
-                local defined_conditional_csname = get_conditional_function_csname(defined_csname, condition)
-                if base_conditional_csname == nil or defined_conditional_csname == nil then  -- we couldn't determine a csname, give up
-                  goto other_statement
-                end
-                local confidence = math.min(argument_specifier_confidence, condition_confidence)
+                local base_conditional_csname = get_conditional_function_csname(base_csname_stem, base_argument_specifiers, condition)
+                local defined_conditional_csname = get_conditional_function_csname(base_csname_stem, argument_specifiers, condition)
+                local confidence = math.min(argument_specifiers.confidence, condition_confidence)
                 if base_conditional_csname ~= defined_conditional_csname then
                   table.insert(defined_csnames, {base_conditional_csname, defined_conditional_csname, confidence})
                 end
               end
             else  -- non-conditional function
+              local defined_csname = replace_argument_specifiers(base_csname_stem, argument_specifiers)
               if base_csname ~= defined_csname then
-                table.insert(defined_csnames, {base_csname, defined_csname, argument_specifier_confidence})
+                table.insert(defined_csnames, {base_csname, defined_csname, argument_specifiers.confidence})
               end
             end
           end
@@ -356,6 +462,7 @@ local function semantic_analysis(pathname, content, issues, results, options)
               -- The following attributes are specific to the type.
               base_csname = effective_base_csname,
               defined_csname = defined_csname,
+              is_private = is_function_private(base_csname),
               is_conditional = is_conditional,
             }
             table.insert(statements, statement)
@@ -385,7 +492,7 @@ local function semantic_analysis(pathname, content, issues, results, options)
             assert(defined_csname ~= nil)
             -- determine the number of parameters of the defined function
             local num_parameters
-            local _, argument_specifiers = parse_expl3_csname(defined_csname)
+            local defined_csname_stem, argument_specifiers = parse_expl3_csname(defined_csname)
             if argument_specifiers ~= nil and lpeg.match(parsers.N_or_n_type_argument_specifiers, argument_specifiers) ~= nil then
               num_parameters = #argument_specifiers
             end
@@ -435,10 +542,7 @@ local function semantic_analysis(pathname, content, issues, results, options)
                 if condition == "p" and is_protected then
                   issues:add("e404", "protected predicate function", byte_range)
                 end
-                local effectively_defined_csname = get_conditional_function_csname(defined_csname, condition)
-                if effectively_defined_csname == nil then  -- we couldn't determine the defined csname, give up
-                  goto other_statement
-                end
+                local effectively_defined_csname = get_conditional_function_csname(defined_csname_stem, argument_specifiers, condition)
                 table.insert(effectively_defined_csnames, {effectively_defined_csname, confidence})
               end
             else  -- non-conditional function
@@ -454,6 +558,7 @@ local function semantic_analysis(pathname, content, issues, results, options)
                 -- The following attributes are specific to the type.
                 subtype = FUNCTION_DEFINITION_DIRECT,
                 maybe_redefinition = maybe_redefinition,
+                is_private = is_function_private(defined_csname),
                 is_global = is_global,
                 defined_csname = effectively_defined_csname,
                 -- The following attributes are specific to the subtype.
@@ -488,6 +593,15 @@ local function semantic_analysis(pathname, content, issues, results, options)
             -- determine all effectively defined csnames and effective base csnames
             local effective_defined_and_base_csnames = {}
             if is_conditional then  -- conditional function
+              -- parse the base and defined csnames
+              local defined_csname_stem, defined_argument_specifiers = parse_expl3_csname(defined_csname)
+              if defined_csname_stem == nil then  -- we couldn't parse the defined csname, give up
+                goto other_statement
+              end
+              local base_csname_stem, base_argument_specifiers = parse_expl3_csname(base_csname)
+              if base_csname_stem == nil then  -- we couldn't parse the base csname, give up
+                goto other_statement
+              end
               -- determine the conditions
               local conditions = parse_conditions(call.arguments[#call.arguments - 1])
               if conditions == nil then  -- we couldn't determine the conditions, give up
@@ -496,11 +610,10 @@ local function semantic_analysis(pathname, content, issues, results, options)
               -- determine the defined and base csnames
               for _, condition_table in ipairs(conditions) do
                 local condition, confidence = table.unpack(condition_table)
-                local effectively_defined_csname = get_conditional_function_csname(defined_csname, condition)
-                local effective_base_csname = get_conditional_function_csname(base_csname, condition)
-                if effectively_defined_csname == nil or effective_base_csname == nil then  -- we couldn't determine a csname, give up
-                  goto other_statement
-                end
+                local effectively_defined_csname
+                  = get_conditional_function_csname(defined_csname_stem, defined_argument_specifiers, condition)
+                local effective_base_csname
+                  = get_conditional_function_csname(base_csname_stem, base_argument_specifiers, condition)
                 table.insert(effective_defined_and_base_csnames, {effectively_defined_csname, effective_base_csname, confidence})
               end
             else  -- non-conditional function
@@ -517,6 +630,7 @@ local function semantic_analysis(pathname, content, issues, results, options)
                 -- The following attributes are specific to the type.
                 subtype = FUNCTION_DEFINITION_INDIRECT,
                 maybe_redefinition = maybe_redefinition,
+                is_private = is_function_private(defined_csname),
                 is_global = is_global,
                 defined_csname = effectively_defined_csname,
                 -- The following attributes are specific to the subtype.
@@ -646,15 +760,12 @@ local function semantic_analysis(pathname, content, issues, results, options)
     end
   end
 
-  -- Determine whether a function is private or public based on its name.
-  local function is_function_private(csname)
-    return csname:sub(1, 2) == "__"
-  end
-
   --- Make a pass over the segments, building up information.
-  local defined_private_functions, defined_private_function_variants = {}, {}
+  local defined_private_functions = {}
+  local defined_private_function_variant_texts, defined_private_function_variant_pattern = {}, parsers.fail
+  local defined_private_function_variant_byte_ranges = {}
   local variant_base_csnames = {}
-  local maybe_defined_csnames, maybe_used_csnames = {}, {}
+  local maybe_used_csnames, maybe_defined_csname_texts, maybe_defined_csname_pattern = {}, {}, parsers.fail
   for segment_number, segment_statements in ipairs(statement_segments) do
     local segment_calls = call_segments[segment_number]
     local segment_tokens, segment_transformed_tokens, map_forward = table.unpack(token_segments[segment_number])
@@ -677,17 +788,41 @@ local function semantic_analysis(pathname, content, issues, results, options)
       if statement.type == FUNCTION_VARIANT_DEFINITION then
         -- Record private function variant defitions.
         maybe_used_csnames[statement.base_csname] = true
-        maybe_defined_csnames[statement.defined_csname] = true
+        if statement.defined_csname.type == TEXT then
+          maybe_defined_csname_texts[statement.defined_csname.payload] = true
+        elseif statement.defined_csname.type == PATTERN then
+          maybe_defined_csname_pattern = (
+            maybe_defined_csname_pattern
+            + statement.defined_csname.payload
+          )
+        else
+          error('Unexpected csname type "' .. statement.defined_csname.type .. '"')
+        end
         table.insert(variant_base_csnames, {statement.base_csname, byte_range})
         if statement.confidence == DEFINITELY then
-          if is_function_private(statement.defined_csname) then
-            table.insert(defined_private_function_variants, {statement.defined_csname, byte_range})
+          if statement.is_private then
+            table.insert(defined_private_function_variant_byte_ranges, byte_range)
+            local defined_private_function_variant = {
+              number = #defined_private_function_variant_byte_ranges,
+              csname = statement.defined_csname
+            }
+            if statement.defined_csname.type == TEXT then
+              table.insert(defined_private_function_variant_texts, defined_private_function_variant)
+            elseif statement.defined_csname.type == PATTERN then
+              defined_private_function_variant_pattern = (
+                defined_private_function_variant_pattern
+                + statement.defined_csname.payload
+                / defined_private_function_variant
+              )
+            else
+              error('Unexpected csname type "' .. statement.defined_csname.type .. '"')
+            end
           end
         end
       elseif statement.type == FUNCTION_DEFINITION then
         -- Record private function defitions.
-        maybe_defined_csnames[statement.defined_csname] = true
-        if statement.confidence == DEFINITELY and is_function_private(statement.defined_csname) then
+        maybe_defined_csname_texts[statement.defined_csname] = true
+        if statement.confidence == DEFINITELY and statement.is_private then
           table.insert(defined_private_functions, {statement.defined_csname, byte_range})
         end
         -- Record base control sequences used as the base in indirect function definitions.
@@ -713,7 +848,7 @@ local function semantic_analysis(pathname, content, issues, results, options)
                           next_token.type == CONTROL_SEQUENCE
                           and lpeg.match(parsers.expl3_function_definition_or_assignment_csname, token.payload) ~= nil
                         ) then
-                      maybe_defined_csnames[next_token.payload] = true
+                      maybe_defined_csname_texts[next_token.payload] = true
                     end
                   end
                 end
@@ -734,6 +869,10 @@ local function semantic_analysis(pathname, content, issues, results, options)
     end
   end
 
+  -- Finalize PEG patterns.
+  maybe_defined_csname_pattern = maybe_defined_csname_pattern * parsers.eof
+  defined_private_function_variant_pattern = defined_private_function_variant_pattern * parsers.eof
+
   --- Report issues apparent from the collected information.
   for _, defined_private_function in ipairs(defined_private_functions) do
     local defined_csname, byte_range = table.unpack(defined_private_function)
@@ -741,15 +880,35 @@ local function semantic_analysis(pathname, content, issues, results, options)
       issues:add('w401', 'unused private function', byte_range)
     end
   end
-  for _, defined_private_function_variant in ipairs(defined_private_function_variants) do
-    local defined_csname, byte_range = table.unpack(defined_private_function_variant)
-    if not maybe_used_csnames[defined_csname] then
+
+  local used_private_function_variants = {}
+  for private_function_variant_number, _ in ipairs(defined_private_function_variant_byte_ranges) do
+    used_private_function_variants[private_function_variant_number] = false
+  end
+  for _, defined_private_function_variant in ipairs(defined_private_function_variant_texts) do
+    assert(defined_private_function_variant.csname.type == TEXT)
+    if maybe_used_csnames[defined_private_function_variant.csname.payload] then
+      used_private_function_variants[defined_private_function_variant.number] = true
+    end
+  end
+  for maybe_used_csname, _ in pairs(maybe_used_csnames) do
+    local defined_private_function_variant = lpeg.match(defined_private_function_variant_pattern, maybe_used_csname)
+    if defined_private_function_variant ~= nil then
+      assert(defined_private_function_variant.csname.type == PATTERN)
+      used_private_function_variants[defined_private_function_variant.number] = true
+    end
+  end
+  for private_function_variant_number, byte_range in ipairs(defined_private_function_variant_byte_ranges) do
+    if not used_private_function_variants[private_function_variant_number] then
       issues:add('w402', 'unused private function variant', byte_range)
     end
   end
+
   for _, variant_base_csname in ipairs(variant_base_csnames) do
     local base_csname, byte_range = table.unpack(variant_base_csname)
-    if not maybe_defined_csnames[base_csname] and lpeg.match(parsers.expl3_maybe_standard_library_csname, base_csname) == nil then
+    if lpeg.match(parsers.expl3_maybe_standard_library_csname, base_csname) == nil
+        and not maybe_defined_csname_texts[base_csname]
+        and not lpeg.match(maybe_defined_csname_pattern, base_csname) then
       issues:add('e405', 'function variant for an undefined function', byte_range)
     end
   end